+++ src/lib/utils.hpp
@@ -212,6 +212,8 @@ void write_BlurEZD (const gridpt grid[], const char outfile[]);
 **************************************************/
 int writeMRCFile(const gridpt data[], const char filename[] );
 int writeSmallMRCFile(const gridpt data[], const char filename[] );
+int writeCCP4File(const gridpt data[], const char filename[] );
+int writeSmallCCP4File(const gridpt data[], const char filename[] );
 void write_output_files(const gridpt grid[],
                         const vossvolvox::OutputSettings& outputs);
 std::unique_ptr<gridpt[]> make_zeroed_grid();